Make sidebar component reusable (#67)
