+++ frontend/src/components/pages/home/index.tsx
@@ -1,13 +1,20 @@
 import Footer from './footer';
 import Header from './header';
 import Introduction from './introduction';
-import Sidebar from './sidebar';
+import Sidebar from '../../common/Sidebar';
 import PhotoServices from './photo-services';
 import Popup from './popup';
 import Products from './products';
 import StockProducts from './stock-products';
+import { useRecoilState, useRecoilValue } from 'recoil';
+import { sidebarOpenState, SidebarType, sidebarTypeState } from '../../../state/sidebar';
+import PriceList from './sidebar/PriceList';
+import Navigation from './sidebar/Navigation';
 
 function Home() {
+    const [sidebarOpen, setSidebarOpen] = useRecoilState(sidebarOpenState);
+    const sidebarType = useRecoilValue(sidebarTypeState);
+
     return (
         <>
             <Header />
@@ -17,7 +24,15 @@ function Home() {
             <StockProducts />
             <Footer />
 
-            <Sidebar />
+            <Sidebar open={sidebarOpen} setOpen={setSidebarOpen}>
+                {sidebarType == SidebarType.Navigation ? (
+                    <Navigation />
+                ) : sidebarType == SidebarType.PriceList ? (
+                    <PriceList />
+                ) : (
+                    <></>
+                )}
+            </Sidebar>
             <Popup />
         </>
     );
